# data/openapispec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Contact Management API By Pavan",
    "description": "API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1"
    }
  ],
  "paths": {
...
    "/contact/create": {
      "post": {
        "summary": "Create a new contact",
        "description": "Creates a new contact for the authenticated user.",
        "operationId": "createContact",

        "security": [
          {
            "CookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,

          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContact"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Contact created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateContactResponseBody"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request (validation error or invalid timezone)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "'Already Exists' if the contact already exists.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "'Already Exists'"
                    }
                  }
                }
              }
            }
          },

          "401": {
            "description": "'Unauthorized' if token verification fails.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "'Unauthorized'"
                    }
                  }
                }
              }
            }
          }
        }
      },

      "put": {
        "summary": "Update an existing contact",
        "description": "Updates an existing contact for the authenticated user.",
        "operationId": "updateContact",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "The name of the contact."
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The email address of the contact."
                  },
                  "phoneNumber": {
                    "type": "string",
                    "description": "The phone number of the contact."
                  },
                  "address": { "type": "string" }
                },
                "required": ["email"]
              }
            }
          }
        },

        "responses": {
          "200": {
            "description": "Contact updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },

          "400": {
            "description": "Bad Request (validation error or invalid timezone)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "additionalProperties": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },

          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "'Contact not found'"
                    }
                  }
                }
              }
            }
          },

          "401": {
            "description": "'Unauthorized' if token verification fails.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "'Unauthorized'"
                    }
                  }
                }
              }
            }
          }
        }
      },

      "delete": {
        "summary": "Soft delete a contact",
        "description": "Marks a contact as deleted for the authenticated user.",
        "operationId": "deleteContact",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "description": "The email of the contact to be deleted."
                  }
                },
                "required": ["email"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contact soft deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contact"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "'Contact not found'"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "'Unauthorized' if token verification fails.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "'Unauthorized'"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "CookieAuth": {
        "type": "apiKey",
        "name": "token",
        "in": "header",
        "description": "The authorization token in the form of `Bearer <token>`"
      }
    },
...
    "schemas": {
      "Contact": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the contact."
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "The email address of the contact."
          },
          "phoneNumber": {
            "type": "string",
            "description": "The phone number of the contact."
          },
          "timezone": {
            "type": "string"
          }
        }
      },
      "CreateContact": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "The name of the contact."
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "The email address of the contact."
          },
          "phoneNumber": {
            "type": "string",
            "description": "The phone number of the contact."
          },
          "address": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          }
        },
        "required": ["email", "name", "phoneNumber", "address", "timezone"]
      },
      "CreateContactResponseBody": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "The unique identifier for the contact."
          },
          "name": {
            "type": "string",
            "description": "The name of the contact."
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "The email address of the contact."
          },
          "phoneNumber": {
            "type": "string",
            "description": "The phone number of the contact."
          },
          "address": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the contact was created."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "The date and time when the contact was created."
          }
        }
      },
...
    },
...
  }
}